partials: support details button in publication links
Add a configuration option to enable showing a details button for each
publication. It may be useful to make a button for the details similar
to the other buttons, in order to give it the same emphasis. Because
this replicates the functionality of the title link, make this optional
(and disabled by default).